+++ extensions/apple-reminders/swift/AppleReminders/Sources/EventKit+Additions.swift
@@ -243,6 +243,25 @@ extension EKReminder {
       }
     }
 
+    var earlyReminderSeconds: Double? = nil
+    if let alarms = self.alarms {
+      for alarm in alarms where !alarm.isLocationAlarm {
+        if alarm.relativeOffset < 0 {
+          earlyReminderSeconds = abs(alarm.relativeOffset)
+          break
+        } else if let alarmDate = alarm.absoluteDate,
+                  let dueDateComponents,
+                  let dueDate = Calendar.current.date(from: dueDateComponents),
+                  alarmDate < dueDate {
+          let diff = dueDate.timeIntervalSince(alarmDate)
+          if diff > 0 {
+            earlyReminderSeconds = diff
+            break
+          }
+        }
+      }
+    }
+
     return Reminder(
       id: self.calendarItemIdentifier,
       openUrl: "x-apple-reminderkit://REMCDReminder/\(self.calendarItemIdentifier)",
@@ -252,7 +271,8 @@ extension EKReminder {
       completionDate: completionDateString, isRecurring: isRecurring,
       recurrenceRule: recurrenceRuleDescription ?? "",
       list: calendar.toStruct(defaultCalendarId: nil), location: location,
-      creationDate: self.creationDate)
+      creationDate: self.creationDate,
+      earlyReminder: earlyReminderSeconds)
   }
 }